[CORDA-1970]: Introduced Caffeine for caching.
