+++ tools/error-codes-server/src/main/kotlin/net/corda/tools/error/codes/server/application/CacheSize.kt
@@ -1,0 +1,8 @@
+package net.corda.tools.error.codes.server.application
+
+internal data class CacheSize(val value: Long) {
+
+    init {
+        require(value > 0)
+    }
+}
